+++ CHANGELOG.md
@@ -25,8 +25,8 @@ ENHANCEMENTS:
 * Error messages for preconditions, postconditions, and custom variable validations are now evaluated as expressions, allowing interpolation of relevant values into the output. ([#30613](https://github.com/hashicorp/terraform/issues/30613))
 * There are some small improvements to the error and warning messages Terraform will emit in the case of invalid provider configuration passing between modules. There are no changes to which situations will produce errors and warnings, but the messages now include additional information intended to clarify what problem Terraform is describing and how to address it. ([#30639](https://github.com/hashicorp/terraform/issues/30639))
 * When running `terraform plan`, only show external changes which may have contributed to the current plan ([#30486](https://github.com/hashicorp/terraform/issues/30486))
-* Add `TF_ORGANIZATION` environment variable fallback for `organization` in the cloud configuration
-* Add `TF_HOSTNAME` environment variable fallback for `hostname` in the cloud configuration
+* Add `TF_CLOUD_ORGANIZATION` environment variable fallback for `organization` in the cloud configuration
+* Add `TF_CLOUD_HOSTNAME` environment variable fallback for `hostname` in the cloud configuration
 * `TF_WORKSPACE` can now be used to configure the `workspaces` attribute in your cloud configuration
 * When running on macOS, Terraform will now use platform APIs to validate certificates presented by TLS (HTTPS) servers. This may change exactly which root certificates Terraform will accept as valid. ([#30768](https://github.com/hashicorp/terraform/issues/30768))
 * The AzureRM Backend now defaults to using MSAL (and Microsoft Graph) rather than ADAL (and Azure Active Directory Graph) for authentication. ([#30891](https://github.com/hashicorp/terraform/issues/30891))